requirements.txt
fastapi
uvicorn[standard]
pillow
ultralytics
torch
opencv-python-headless
numpy
python-multipart